requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
pydantic
openai
slack-sdk
jira
reportlab
ollama
python-dotenv
python-multipart
pandas
gunicorn